lib: cgen: packet: move set codegen from set.c to packet.c
