+++ src/libbpfilter/cgen/matcher/set.c
@@ -5,9 +5,7 @@
 
 #include "cgen/matcher/set.h"
 
-#include <bpfilter/chain.h>
 #include <bpfilter/matcher.h>
-#include <bpfilter/set.h>
 
 #include "cgen/program.h"
 #include "cgen/stub.h"
@@ -48,58 +46,4 @@ int bf_set_generate_trie_lookup(struct bf_program *program,
         return r;
 
     return bf_set_generate_map_lookup(program, matcher, BF_PROG_SCR_OFF(4));
-}
-
-int bf_matcher_generate_set(struct bf_program *program,
-                            const struct bf_matcher *matcher)
-{
-    assert(program);
-    assert(matcher);
-
-    const struct bf_set *set =
-        bf_chain_get_set_for_matcher(program->runtime.chain, matcher);
-    size_t offset = 0;
-    int r;
-
-    if (!set) {
-        return bf_err_r(-ENOENT, "set #%u not found in %s",
-                        *(uint32_t *)bf_matcher_payload(matcher),
-                        program->runtime.chain->name);
-    }
-
-    if (set->use_trie) {
-        const struct bf_matcher_meta *meta = bf_matcher_get_meta(set->key[0]);
-
-        r = bf_stub_load_header(program, meta, BPF_REG_6);
-        if (r)
-            return bf_err_r(r, "failed to load protocol header into BPF_REG_6");
-
-        return bf_set_generate_trie_lookup(
-            program, matcher, meta->hdr_payload_offset, meta->hdr_payload_size);
-    }
-
-    // Generate the bytecode to build the set key
-    for (size_t i = 0; i < set->n_comps; ++i) {
-        enum bf_matcher_type type = set->key[i];
-        const struct bf_matcher_meta *meta = bf_matcher_get_meta(type);
-
-        if (!meta) {
-            return bf_err_r(-ENOENT, "meta for '%s' not found",
-                            bf_matcher_type_to_str(type));
-        }
-
-        r = bf_stub_load_header(program, meta, BPF_REG_6);
-        if (r)
-            return bf_err_r(r, "failed to load protocol header into BPF_REG_6");
-
-        r = bf_stub_stx_payload(program, meta, offset);
-        if (r) {
-            return bf_err_r(r,
-                            "failed to generate bytecode to load packet data");
-        }
-
-        offset += meta->hdr_payload_size;
-    }
-
-    return bf_set_generate_map_lookup(program, matcher, 0);
 }
